pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "ultron-bot"
version = "0.2.31"
description = "Discord bot connecting Redmine with OpenAI-compatible LLMs"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
dependencies = [
  "colorlog>=6.8.0",
  "discord.py>=2.3.0",
  "httpx>=0.27.0",
  "openai>=1.40.0",
  "python-dotenv>=1.0.0",
  "PyYAML>=6.0.1",
]

[project.scripts]
ultron = "ultron.__main__:main"

[project.optional-dependencies]
dev = ["pytest>=8.0.0"]
wizard = ["questionary>=2.0.0"]

[tool.setuptools.packages.find]
where = ["."]
include = ["ultron*"]
